feat: add update_json function
This function takes two string _assumed_ to be valid json and
then gives their "concatenated" format. The implementation is
very simplistic and takes a lot of assumptions this should be
improved in the future. This function also assumes that we would
read teh whole JSON file which is less than ideal since it would
become a huge JSON file in the future and it would be useless to
store that data

Signed-off-by: Akshat Agarwal <humancalico@disroot.org>