feat(webhooks): create a basic server
This commit creates a basic server that logs the commit hash of the
commit when the check_suite is requested or run. This only logs the it
when check suite has been completed. Although I am not sure if mulitple
check suites are possible in a single commit or merge

Signed-off-by: Akshat Agarwal <humancalico@disroot.org>